Extendible Hash
Implemented an extendible hash. My code is in extendiblehash.cpp and
extendiblehash.h. The rest of the files are the instructor's which are
used in conjunction with my files in order to run it.

Create extendablefile.cpp generates a test file for our main method
(Hash-#-#-#-#)
Hash.data files are generated by create extendiblefile and are used to
test to program.
First number determines the max size of the hash
Second number determines the leaf size
Third number determines how many bits we start with for the hash index
Last number generates the seed.

b denotes when we have the wrong amount of indices.
i denotes an insertion
d denotes a deletion
f denotes a search for a particular integer, returns -1 if not found.